退积分、退余额

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/order/virtual/MemberCardRefundParam.java
@@ -16,5 +16,5 @@ public class MemberCardRefundParam {
     /** 现金 **/
     private Double money;
     /** 积分 **/
-    private Double score;
+    private Integer score;
 }
